Moved Report to panels
