+++ src/components/Controllers/File/Panels/Report.tsx
@@ -1,14 +1,8 @@
 import * as React from 'react'
-import Box from '@mui/material/Box'
-import Report from '../../../Parts/Report'
+import Report from '../../../Parts/Panels/Report'
 import { useStore } from '../store'
 
 export default function ReportPanel() {
   const report = useStore((state) => state.report)
-  if (!report) return null
-  return (
-    <Box sx={{ paddingX: 2 }}>
-      <Report report={report} />
-    </Box>
-  )
+  return <Report report={report} />
 }
